chore: spacing + remove unnecessary files
