--- kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/classes/labelled/Labelled.kt
@@ -1,5 +1,0 @@
-package xyz.lbres.kotlinutils.classes.labelled
-
-import xyz.lbres.kotlinutils.general.labelled.Labelled
-
-typealias Labelled<T> = Labelled<T>
