+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/classes/labelled/references.kt
@@ -1,0 +1,12 @@
+// ktlint-disable filename
+
+package xyz.lbres.kotlinutils.classes.labelled
+
+import xyz.lbres.kotlinutils.general.labelled.Labelled
+
+/**
+ * Code for the Labelled class has been moved to the general/labelled package.
+ * This file exists only to avoid breaking existing functionality that uses the class at this path.
+ */
+
+typealias Labelled<T> = Labelled<T>
