--- kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/classes/multiset/MutableMultiSet.kt
@@ -1,9 +1,0 @@
-package xyz.lbres.kotlinutils.classes.multiset
-
-import xyz.lbres.kotlinutils.set.multiset.MutableMultiSet
-
-/**
- * Interface for set that allows multiple occurrences of a value.
- * The interface supports read/write access to the values.
- */
-typealias MutableMultiSet<E> = MutableMultiSet<E>
